+++ frontend/src/metabase-lib/query/clauses/define.ts
