+++ frontend/src/metabase-lib/query/clauses/index.ts
@@ -1,0 +1,2 @@
+export * from "./clauses";
+export * from "./types";
