+++ frontend/src/metabase-lib/query/types.ts
@@ -1,5 +1,4 @@
 /* eslint-disable @typescript-eslint/no-unused-vars -- used for types */
-import type { DefinedClauseName } from "metabase/querying/expressions";
 import type {
   CardId,
   DatabaseId,
@@ -30,6 +29,7 @@ import type {
   TimeFilterOperator,
 } from "../common";
 
+import type { DefinedClauseName } from "./clauses";
 import type { ColumnExtractionTag } from "./extractions";
 
 /**
