+++ frontend/src/metabase/querying/expressions/help-text.unit.spec.ts
@@ -1,9 +1,9 @@
 import { createMockMetadata } from "__support__/metadata";
 import { checkNotNull } from "metabase/utils/types";
+import * as Lib from "metabase-lib";
 import type { Database } from "metabase-types/api";
 import { createMockDatabase } from "metabase-types/api/mocks/database";
 
-import { MBQL_CLAUSES } from "./config";
 import { formatExpressionParts } from "./formatter";
 import { getHelpText } from "./help-text";
 
@@ -150,7 +150,7 @@ describe("getHelpText", () => {
   });
 
   it("all help texts can be formatted", async () => {
-    for (const name in MBQL_CLAUSES) {
+    for (const name in Lib.MBQL_CLAUSES) {
       const { database } = setup();
       const helpText = getHelpText(name, database, reportTimezone);
       if (!helpText) {
